requirement.txt
cx-Freeze
mitmproxy==11.0.2
psutil==7.2.2
websocket-client==1.9.0
